Chat header + entities edit

+++ server/src/main/java/com/dzalex/skillshuffle/repositories/ChatMemberRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ChatMemberRepository extends JpaRepository<ChatMember, Long> {
+public interface ChatMemberRepository extends JpaRepository<ChatMember, Integer> {
     List<ChatMember> findAllByChatId(Integer chatId);
 }
