+++ server/src/main/java/com/dzalex/skillshuffle/repositories/ChatRepository.java
@@ -3,6 +3,6 @@ package com.dzalex.skillshuffle.repositories;
 import com.dzalex.skillshuffle.entities.Chat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatRepository extends JpaRepository<Chat, Long> {
+public interface ChatRepository extends JpaRepository<Chat, Integer> {
     Chat findChatById(Integer id);
 }
