+++ server/src/main/java/com/dzalex/skillshuffle/repositories/MessageRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MessageRepository  extends JpaRepository<ChatMessage, Long> {
+public interface MessageRepository  extends JpaRepository<ChatMessage, Integer> {
     ChatMessage findMessageById(Integer id);
     List<ChatMessage> findMessagesByChatId(Integer chatId);
 }
